feat(产检预约):产检预约预约逻辑

+++ ih_patient/src/main/java/com/keydom/ih_patient/activity/pregnancy/view/PregnancyDetailView.java
@@ -39,7 +39,7 @@ public interface PregnancyDetailView extends BaseView {
 
     int getAppointType();
 
-    int getPrenatalProjectId();
+    long getPrenatalProjectId();
 
     String getTimeInterval();
 }
